fixed range annotations

+++ Shared/ViewModels/PagingVM.cs
@@ -7,7 +7,7 @@ namespace Shared.ViewModels
 {
     public class PagingVM
     {
-        [Range(minimum: 0, maximum: double.MaxValue, ErrorMessage ="The field {0} must be greater than {1}")]
+        [Range(minimum: 0, maximum: int.MaxValue, ErrorMessage ="The field {0} must be greater than {1}")]
         public int PageNumber { get; set; }
         [Range(minimum: 0, maximum: int.MaxValue, ErrorMessage = "The field {0} must be greater than {1}")]
         public int PageSize { get; set; }
